fix/fixValidationsClass | refactoring all validations classes
