+++ src/main/java/com/challenge/hotel_california/validatorRefactor/bookingsCreateValidation/IValidatorBookings.java
@@ -1,0 +1,10 @@
+package com.challenge.hotel_california.validatorRefactor.bookingsCreateValidation;
+
+import com.challenge.hotel_california.DTOs.BookingEntryDTO;
+import com.challenge.hotel_california.model.Customer;
+import com.challenge.hotel_california.model.Room;
+
+public interface IValidatorBookings {
+    void verifyValidatorsBookings(BookingEntryDTO bookingEntryDTO, Room room, Customer customer);
+
+}
