+++ src/main/java/com/challenge/hotel_california/validatorRefactor/bookingsDeleteValidation/IValidatorBookingsDelete.java
@@ -1,0 +1,7 @@
+package com.challenge.hotel_california.validatorRefactor.bookingsDeleteValidation;
+
+import com.challenge.hotel_california.model.Booking;
+
+public interface IValidatorBookingsDelete {
+    void verifyBookingsDeleteValidators(Long id, Booking bookingFound);
+}
