+++ src/main/java/com/challenge/hotel_california/validatorRefactor/bookingsDeleteValidation/validatorTimeBookingIsBefore48Hour.java
@@ -1,34 +1,21 @@
-package com.challenge.hotel_california.validatorRefactor;
+package com.challenge.hotel_california.validatorRefactor.bookingsDeleteValidation;
 
 import com.challenge.hotel_california.enums.BookingStatus;
 import com.challenge.hotel_california.enums.RoomStatus;
 import com.challenge.hotel_california.exceptions.BookingStatusException;
-import com.challenge.hotel_california.exceptions.BookingsNotFoundException;
 import com.challenge.hotel_california.model.Booking;
-import com.challenge.hotel_california.repository.BookingRepository;
 import com.challenge.hotel_california.service.BookingService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
+
 @Component
-public class validatorBookingsDelete implements IValidatorBookingsDeleteAndCheckout {
-    @Autowired
-    private BookingRepository bookingRepository;
+public class validatorTimeBookingIsBefore48Hour implements IValidatorBookingsDelete {
 
     @Override
-    public void verifyBookingsValidators(Long id, Booking bookingFound) {
-        if (!bookingRepository.existsById(id)) {
-            throw new BookingsNotFoundException("Booking not found");
-        }
+    public void verifyBookingsDeleteValidators(Long id, Booking bookingFound) {
         var bookingStatusFound = bookingFound.getStatus();
-        if (bookingStatusFound.equals(BookingStatus.CHECKED_IN)) {
-            throw new BookingStatusException("This Reservation cannot be cancelled case its made a Check in!");
-        }
-        if (bookingStatusFound.equals(BookingStatus.COMPLETED) || bookingStatusFound.equals(BookingStatus.CANCELLED)) {
-            throw new BookingStatusException("Cannot cancel a completed or cancelled reservation!");
-        }
         if (bookingStatusFound.equals(BookingStatus.CONFIRMED)) {
             var checkInCurrentDate = bookingFound.getCheckInDate();
             var cancelDate = LocalDateTime.now();
