+++ src/main/java/com/challenge/hotel_california/validatorRefactor/bookingsUpdateCheckoutValidation/IValidatorBookingsCheckout.java
@@ -1,0 +1,7 @@
+package com.challenge.hotel_california.validatorRefactor.bookingsUpdateCheckoutValidation;
+
+import com.challenge.hotel_california.model.Booking;
+
+public interface IValidatorBookingsCheckout {
+    void verifyBookingsCheckoutValidators(Long id, Booking bookingFound);
+}
