+++ src/main/java/com/challenge/hotel_california/validatorRefactor/bookingsUpdateCheckoutValidation/ValidateCheckoutDateValidationIsAfterTheCheckInDate.java
@@ -1,12 +1,8 @@
-package com.challenge.hotel_california.validatorRefactor;
+package com.challenge.hotel_california.validatorRefactor.bookingsUpdateCheckoutValidation;
 
-import com.challenge.hotel_california.enums.BookingStatus;
 import com.challenge.hotel_california.exceptions.BookingCheckInDateNotBeforeException;
-import com.challenge.hotel_california.exceptions.BookingsNotFoundException;
-import com.challenge.hotel_california.exceptions.RoomNotAvailableException;
 import com.challenge.hotel_california.model.Booking;
 import com.challenge.hotel_california.repository.BookingRepository;
-import com.challenge.hotel_california.validatorRefactor.IValidatorBookingsDeleteAndCheckout;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -14,21 +10,14 @@ import java.math.BigDecimal;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
+
 @Component
-public class ValidatorBookingsUpdateCheckOut implements IValidatorBookingsDeleteAndCheckout {
+class ValidateCheckoutDateValidationIsAfterTheCheckInDate implements IValidatorBookingsCheckout {
     @Autowired
     private BookingRepository bookingRepository;
 
     @Override
-    public void verifyBookingsValidators(Long id, Booking bookingFound) {
-        if (!bookingRepository.existsById(id)) {
-            throw new BookingsNotFoundException("No Booking was found");
-
-        }
-        if (bookingFound.getStatus().equals(BookingStatus.COMPLETED) || bookingFound.getStatus().equals(BookingStatus.CANCELLED)) {
-            throw new RoomNotAvailableException("Cannot cancel a completed or cancelled reservation!");
-        }
-
+    public void verifyBookingsCheckoutValidators(Long id, Booking bookingFound) {
         LocalTime checkInTime = LocalTime.of(14, 0);  // 14:00
         LocalTime checkOutTime = LocalTime.of(8, 0);  // 08:00
