+++ src/main/java/com/challenge/hotel_california/validatorRefactor/bookingsUpdateValidation/IValidatorBookingsUpdate.java
@@ -1,18 +1,11 @@
-package com.challenge.hotel_california.validatorRefactor;
+package com.challenge.hotel_california.validatorRefactor.bookingsUpdateValidation;
 
-import com.challenge.hotel_california.DTOs.BookingEntryDTO;
 import com.challenge.hotel_california.DTOs.BookingUpdateEntryDTO;
 import com.challenge.hotel_california.model.Booking;
 import com.challenge.hotel_california.model.Customer;
 import com.challenge.hotel_california.model.Room;
 
-public interface IValidatorBookings {
-    void verifyValidatorsBookings(BookingEntryDTO bookingEntryDTO, Room room, Customer customer);
-
+public interface IValidatorBookingsUpdate {
     void verifyBookingsUpdateValidators(BookingUpdateEntryDTO bookingUpdateEntryDTO, Room roomFound,
                                         Booking bookingFound, Customer customerFound, Long id);
-
-
-
-
 }
