+++ src/main/java/com/challenge/hotel_california/validatorRefactor/bookingsUpdateValidation/ValidateIfRoomAlreadyBooked.java
@@ -1,0 +1,30 @@
+package com.challenge.hotel_california.validatorRefactor.bookingsUpdateValidation;
+
+import com.challenge.hotel_california.DTOs.BookingUpdateEntryDTO;
+import com.challenge.hotel_california.enums.RoomStatus;
+import com.challenge.hotel_california.exceptions.BookingsExistsException;
+import com.challenge.hotel_california.model.Booking;
+import com.challenge.hotel_california.model.Customer;
+import com.challenge.hotel_california.model.Room;
+import com.challenge.hotel_california.repository.BookingRepository;
+import com.challenge.hotel_california.repository.RoomRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ValidateIfRoomAlreadyBooked implements IValidatorBookingsUpdate {
+
+    @Autowired
+    private RoomRepository roomRepository;
+    @Autowired
+    private BookingRepository bookingRepository;
+
+    @Override
+    public void verifyBookingsUpdateValidators(BookingUpdateEntryDTO bookingUpdateEntryDTO, Room roomFound,
+                                               Booking bookingFound, Customer customerFound, Long id) {
+
+        if (roomFound.getStatus().equals(RoomStatus.BOOKED) && !roomFound.getId().equals(bookingFound.getRoom().getId())) {
+            throw new BookingsExistsException("This room is already booked!");
+        }
+    }
+}
