+++ src/main/java/com/challenge/hotel_california/validatorRefactor/rooms/ValidateIfBookingsAreActive.java
@@ -1,4 +1,4 @@
-package com.challenge.hotel_california.validatorRefactor;
+package com.challenge.hotel_california.validatorRefactor.rooms;
 
 import com.challenge.hotel_california.DTOs.RoomEntryUpdateDTO;
 import com.challenge.hotel_california.enums.BookingStatus;
@@ -14,28 +14,15 @@ import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Component
-public class ValidatorsRoomsHandler implements IValidatorRooms {
-    @Autowired
-    private RoomRepository roomRepository;
+public class ValidateIfBookingsAreActive implements IValidatorRooms {
+
     @Autowired
     private BookingRepository bookingRepository;
 
     @Override
-    public void verifyRoomUpdateValidators(Long id, Room room, RoomEntryUpdateDTO roomEntryUpdateDTO) {
-        if (!roomRepository.existsById(id)) {
-            throw new RoomNotFoundException("Room " + id + " Not Found in Hotel California Database");
-        }
-        if (!room.getId().equals(roomEntryUpdateDTO.id())) {
-            throw new RoomNotFoundException("Room " + id + " not the same of ID: " + roomEntryUpdateDTO.id());
-        }
-        List<Room> roomList = roomRepository.findAll();
-        var foundRoomNumber = roomList.stream().filter(r -> r.getNumber().equals(roomEntryUpdateDTO.number())).collect(Collectors.toList());
-        if (!room.getNumber().equals(roomEntryUpdateDTO.number()) && !foundRoomNumber.isEmpty()) {
-            throw new NumberRoomFoundException("Room " + roomEntryUpdateDTO.number() + " already Exists");
-        }
+    public void verifyRoomUpdateValidators(Long id, RoomEntryUpdateDTO roomEntryUpdateDTO) {
         List<BookingStatus> optionsStatusBooking = Arrays.asList(BookingStatus.CANCELLED, BookingStatus.COMPLETED);
         List<Booking> bookingsRoom = bookingRepository.getBookingsById(id, optionsStatusBooking);
         if (!bookingsRoom.isEmpty()) {
